+++ src/app/components/_routes/home/home.component.ts
@@ -1,4 +1,5 @@
 import {Component, OnInit} from '@angular/core';
+import {Title} from '@angular/platform-browser';
 
 @Component({
   selector: 'app-home',
@@ -6,5 +7,8 @@ import {Component, OnInit} from '@angular/core';
   styleUrls: ['./home.component.scss'],
 })
 export class HomeComponent implements OnInit {
-  ngOnInit(): void {}
+  ngOnInit(): void {
+    this.titleService.setTitle('Home - Stalker');
+  }
+  constructor(private titleService: Title) {}
 }